+++ src/s21_math.h
@@ -15,7 +15,7 @@ int s21_abs(int x);
 long double s21_ceil(double x);
 long double s21_exp(double x);
 long double s21_fabs(double x);
-long double s21_factorial(long long int x);
+long long int s21_factorial(int x);
 long double s21_floor(double x);
 long double s21_fmod(double x, double y);
 long double s21_log(double x);
